Move and fix path
