+++ superset-frontend/cypress-base/cypress/e2e/applitools/chartlist.test.ts
