+++ superset-frontend/cypress-base/cypress/e2e/applitools/sqllab.test.ts
